refactor: rename AboutUsNavButton into GenericPagesComponents
