--- src/components/GenericPagesComponents/AboutUsNavButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AboutUsNavButton";
